requirements/common.txt
django-extensions==1.5.5
six==1.9.0
django-forms-bootstrap==3.0.1
Pillow==2.9.0
